cpu/12l1r: Move SXTH, UXTH, SXTB, UXTB to proper file category
