+++ src/emu/cpu/CMakeLists.txt
@@ -17,6 +17,7 @@ set(SOURCE_12L1R
         include/cpu/12l1r/visit_session.h
         src/12l1r/translate/branch.cpp
         src/12l1r/translate/data_processing.cpp
+        src/12l1r/translate/extension.cpp
         src/12l1r/translate/interrupt.cpp
         src/12l1r/translate/load_store.cpp
         src/12l1r/translate/multiply.cpp
