+++ src/emu/cpu/src/12l1r/translate/data_processing.cpp
@@ -1830,48 +1830,4 @@ namespace eka2l1::arm::r12l1 {
 
         return true;
     }
-
-    bool thumb_translate_visitor::thumb16_SXTH(reg_index m, reg_index d) {
-        common::armgen::arm_reg dest_real = reg_index_to_gpr(d);
-        common::armgen::arm_reg source_real = reg_index_to_gpr(m);
-
-        common::armgen::arm_reg dest_mapped = reg_supplier_.map(dest_real, ALLOCATE_FLAG_DIRTY);
-        common::armgen::arm_reg source_mapped = reg_supplier_.map(source_real, 0);
-
-        big_block_->SXTH(dest_mapped, source_mapped);
-        return true;
-    }
-
-    bool thumb_translate_visitor::thumb16_SXTB(reg_index m, reg_index d) {
-        common::armgen::arm_reg dest_real = reg_index_to_gpr(d);
-        common::armgen::arm_reg source_real = reg_index_to_gpr(m);
-
-        common::armgen::arm_reg dest_mapped = reg_supplier_.map(dest_real, ALLOCATE_FLAG_DIRTY);
-        common::armgen::arm_reg source_mapped = reg_supplier_.map(source_real, 0);
-
-        big_block_->SXTB(dest_mapped, source_mapped);
-        return true;
-    }
-
-    bool thumb_translate_visitor::thumb16_UXTH(reg_index m, reg_index d) {
-        common::armgen::arm_reg dest_real = reg_index_to_gpr(d);
-        common::armgen::arm_reg source_real = reg_index_to_gpr(m);
-
-        common::armgen::arm_reg dest_mapped = reg_supplier_.map(dest_real, ALLOCATE_FLAG_DIRTY);
-        common::armgen::arm_reg source_mapped = reg_supplier_.map(source_real, 0);
-
-        big_block_->UXTH(dest_mapped, source_mapped);
-        return true;
-    }
-
-    bool thumb_translate_visitor::thumb16_UXTB(reg_index m, reg_index d) {
-        common::armgen::arm_reg dest_real = reg_index_to_gpr(d);
-        common::armgen::arm_reg source_real = reg_index_to_gpr(m);
-
-        common::armgen::arm_reg dest_mapped = reg_supplier_.map(dest_real, ALLOCATE_FLAG_DIRTY);
-        common::armgen::arm_reg source_mapped = reg_supplier_.map(source_real, 0);
-
-        big_block_->UXTB(dest_mapped, source_mapped);
-        return true;
-    }
 }
